+++ Services/Image/V2/ImageClient.cs
@@ -62,6 +62,38 @@ namespace HuaweiCloud.SDK.Image.V2
         }
         
         /// <summary>
+        /// 创建图像重构任务
+        ///
+        /// Create Task
+        /// 
+        /// Please refer to HUAWEI cloud API Explorer for details.
+        /// </summary>
+        public CreateImageVariationTaskResponse CreateImageVariationTask(CreateImageVariationTaskRequest createImageVariationTaskRequest)
+        {
+            Dictionary<string, string> urlParam = new Dictionary<string, string>();
+            string urlPath = HttpUtils.AddUrlPath("/v2/{project_id}/image/image-variation/tasks",urlParam);
+            SdkRequest request = HttpUtils.InitSdkRequest(urlPath, "application/json;charset=UTF-8", createImageVariationTaskRequest);
+            HttpResponseMessage response = DoHttpRequestSync("POST",request);
+            return JsonUtils.DeSerialize<CreateImageVariationTaskResponse>(response);
+        }
+        
+        /// <summary>
+        /// 创建文生图任务
+        ///
+        /// Create Task
+        /// 
+        /// Please refer to HUAWEI cloud API Explorer for details.
+        /// </summary>
+        public CreateTextToImageTaskResponse CreateTextToImageTask(CreateTextToImageTaskRequest createTextToImageTaskRequest)
+        {
+            Dictionary<string, string> urlParam = new Dictionary<string, string>();
+            string urlPath = HttpUtils.AddUrlPath("/v2/{project_id}/image/text-to-image/tasks",urlParam);
+            SdkRequest request = HttpUtils.InitSdkRequest(urlPath, "application/json;charset=UTF-8", createTextToImageTaskRequest);
+            HttpResponseMessage response = DoHttpRequestSync("POST",request);
+            return JsonUtils.DeSerialize<CreateTextToImageTaskResponse>(response);
+        }
+        
+        /// <summary>
         /// 创建任务
         ///
         /// 创建视频封面任务
@@ -286,38 +318,6 @@ namespace HuaweiCloud.SDK.Image.V2
         }
         
         /// <summary>
-        /// 智能设计图像滤镜
-        ///
-        /// 智能设计图像滤镜服务
-        /// 
-        /// Please refer to HUAWEI cloud API Explorer for details.
-        /// </summary>
-        public RunImageWisedesignColorfilterResponse RunImageWisedesignColorfilter(RunImageWisedesignColorfilterRequest runImageWisedesignColorfilterRequest)
-        {
-            Dictionary<string, string> urlParam = new Dictionary<string, string>();
-            string urlPath = HttpUtils.AddUrlPath("/v2/{project_id}/image/image-wisedesign-colorfilter",urlParam);
-            SdkRequest request = HttpUtils.InitSdkRequest(urlPath, "application/json;charset=UTF-8", runImageWisedesignColorfilterRequest);
-            HttpResponseMessage response = DoHttpRequestSync("POST",request);
-            return JsonUtils.DeSerialize<RunImageWisedesignColorfilterResponse>(response);
-        }
-        
-        /// <summary>
-        /// 智能设计图像合图
-        ///
-        /// 智能设计图像合图服务
-        /// 
-        /// Please refer to HUAWEI cloud API Explorer for details.
-        /// </summary>
-        public RunImageWisedesignCombineResponse RunImageWisedesignCombine(RunImageWisedesignCombineRequest runImageWisedesignCombineRequest)
-        {
-            Dictionary<string, string> urlParam = new Dictionary<string, string>();
-            string urlPath = HttpUtils.AddUrlPath("/v2/{project_id}/image/image-wisedesign-combine",urlParam);
-            SdkRequest request = HttpUtils.InitSdkRequest(urlPath, "application/json;charset=UTF-8", runImageWisedesignCombineRequest);
-            HttpResponseMessage response = DoHttpRequestSync("POST",request);
-            return JsonUtils.DeSerialize<RunImageWisedesignCombineResponse>(response);
-        }
-        
-        /// <summary>
         /// 智能设计图像裁剪
         ///
         /// 智能设计图像裁剪服务
@@ -430,6 +430,40 @@ namespace HuaweiCloud.SDK.Image.V2
             SdkRequest request = HttpUtils.InitSdkRequest(urlPath, "application/json", showImageTranslateTaskRequest);
             HttpResponseMessage response = DoHttpRequestSync("GET",request);
             return JsonUtils.DeSerialize<ShowImageTranslateTaskResponse>(response);
+        }
+        
+        /// <summary>
+        /// show task 查询任务信息
+        ///
+        /// show task
+        /// 
+        /// Please refer to HUAWEI cloud API Explorer for details.
+        /// </summary>
+        public ShowImageVariationTaskResponse ShowImageVariationTask(ShowImageVariationTaskRequest showImageVariationTaskRequest)
+        {
+            Dictionary<string, string> urlParam = new Dictionary<string, string>();
+            urlParam.Add("task_id" , showImageVariationTaskRequest.TaskId.ToString());
+            string urlPath = HttpUtils.AddUrlPath("/v2/{project_id}/image/image-variation/tasks/{task_id}",urlParam);
+            SdkRequest request = HttpUtils.InitSdkRequest(urlPath, "application/json", showImageVariationTaskRequest);
+            HttpResponseMessage response = DoHttpRequestSync("GET",request);
+            return JsonUtils.DeSerialize<ShowImageVariationTaskResponse>(response);
+        }
+        
+        /// <summary>
+        /// show task 查询任务信息
+        ///
+        /// show task
+        /// 
+        /// Please refer to HUAWEI cloud API Explorer for details.
+        /// </summary>
+        public ShowTextToImageTaskResponse ShowTextToImageTask(ShowTextToImageTaskRequest showTextToImageTaskRequest)
+        {
+            Dictionary<string, string> urlParam = new Dictionary<string, string>();
+            urlParam.Add("task_id" , showTextToImageTaskRequest.TaskId.ToString());
+            string urlPath = HttpUtils.AddUrlPath("/v2/{project_id}/image/text-to-image/tasks/{task_id}",urlParam);
+            SdkRequest request = HttpUtils.InitSdkRequest(urlPath, "application/json", showTextToImageTaskRequest);
+            HttpResponseMessage response = DoHttpRequestSync("GET",request);
+            return JsonUtils.DeSerialize<ShowTextToImageTaskResponse>(response);
         }
         
         /// <summary>