+++ Services/Image/V2/Model/CreateTextToImageTaskRequestBody.cs
@@ -1,0 +1,105 @@
+using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Linq;
+using System.Runtime.Serialization;
+
+using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+using HuaweiCloud.SDK.Core;
+
+namespace HuaweiCloud.SDK.Image.V2.Model
+{
+    /// <summary>
+    /// 
+    /// </summary>
+    public class CreateTextToImageTaskRequestBody 
+    {
+
+        /// <summary>
+        /// 
+        /// </summary>
+        [JsonProperty("output", NullValueHandling = NullValueHandling.Ignore)]
+        public TaskOutput Output { get; set; }
+
+        /// <summary>
+        /// 
+        /// </summary>
+        [JsonProperty("callback", NullValueHandling = NullValueHandling.Ignore)]
+        public TaskCallback Callback { get; set; }
+
+        /// <summary>
+        /// 
+        /// </summary>
+        [JsonProperty("config", NullValueHandling = NullValueHandling.Ignore)]
+        public TextToImageTaskConfig Config { get; set; }
+
+
+
+        /// <summary>
+        /// Get the string
+        /// </summary>
+        public override string ToString()
+        {
+            var sb = new StringBuilder();
+            sb.Append("class CreateTextToImageTaskRequestBody {\n");
+            sb.Append("  output: ").Append(Output).Append("\n");
+            sb.Append("  callback: ").Append(Callback).Append("\n");
+            sb.Append("  config: ").Append(Config).Append("\n");
+            sb.Append("}\n");
+            return sb.ToString();
+        }
+
+        /// <summary>
+        /// Returns true if objects are equal
+        /// </summary>
+        public override bool Equals(object input)
+        {
+            return this.Equals(input as CreateTextToImageTaskRequestBody);
+        }
+
+        /// <summary>
+        /// Returns true if objects are equal
+        /// </summary>
+        public bool Equals(CreateTextToImageTaskRequestBody input)
+        {
+            if (input == null)
+                return false;
+
+            return 
+                (
+                    this.Output == input.Output ||
+                    (this.Output != null &&
+                    this.Output.Equals(input.Output))
+                ) && 
+                (
+                    this.Callback == input.Callback ||
+                    (this.Callback != null &&
+                    this.Callback.Equals(input.Callback))
+                ) && 
+                (
+                    this.Config == input.Config ||
+                    (this.Config != null &&
+                    this.Config.Equals(input.Config))
+                );
+        }
+
+        /// <summary>
+        /// Get hash code
+        /// </summary>
+        public override int GetHashCode()
+        {
+            unchecked // Overflow is fine, just wrap
+            {
+                int hashCode = 41;
+                if (this.Output != null)
+                    hashCode = hashCode * 59 + this.Output.GetHashCode();
+                if (this.Callback != null)
+                    hashCode = hashCode * 59 + this.Callback.GetHashCode();
+                if (this.Config != null)
+                    hashCode = hashCode * 59 + this.Config.GetHashCode();
+                return hashCode;
+            }
+        }
+    }
+}