+++ Services/Image/V2/Model/ImageVariationInference.cs
@@ -1,0 +1,91 @@
+using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Linq;
+using System.Runtime.Serialization;
+
+using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+using HuaweiCloud.SDK.Core;
+
+namespace HuaweiCloud.SDK.Image.V2.Model
+{
+    /// <summary>
+    /// 
+    /// </summary>
+    public class ImageVariationInference 
+    {
+
+        /// <summary>
+        /// 随机数种子
+        /// </summary>
+        [JsonProperty("seed", NullValueHandling = NullValueHandling.Ignore)]
+        public int? Seed { get; set; }
+
+        /// <summary>
+        /// 生成图片分辨率，限定字符串\&quot;512*768\&quot;,\&quot;768*512\&quot;,\&quot;512*512\&quot;,\&quot;1024*768\&quot;,\&quot;768*1024\&quot;，默认\&quot;512*512\&quot;
+        /// </summary>
+        [JsonProperty("resolution", NullValueHandling = NullValueHandling.Ignore)]
+        public string Resolution { get; set; }
+
+
+
+        /// <summary>
+        /// Get the string
+        /// </summary>
+        public override string ToString()
+        {
+            var sb = new StringBuilder();
+            sb.Append("class ImageVariationInference {\n");
+            sb.Append("  seed: ").Append(Seed).Append("\n");
+            sb.Append("  resolution: ").Append(Resolution).Append("\n");
+            sb.Append("}\n");
+            return sb.ToString();
+        }
+
+        /// <summary>
+        /// Returns true if objects are equal
+        /// </summary>
+        public override bool Equals(object input)
+        {
+            return this.Equals(input as ImageVariationInference);
+        }
+
+        /// <summary>
+        /// Returns true if objects are equal
+        /// </summary>
+        public bool Equals(ImageVariationInference input)
+        {
+            if (input == null)
+                return false;
+
+            return 
+                (
+                    this.Seed == input.Seed ||
+                    (this.Seed != null &&
+                    this.Seed.Equals(input.Seed))
+                ) && 
+                (
+                    this.Resolution == input.Resolution ||
+                    (this.Resolution != null &&
+                    this.Resolution.Equals(input.Resolution))
+                );
+        }
+
+        /// <summary>
+        /// Get hash code
+        /// </summary>
+        public override int GetHashCode()
+        {
+            unchecked // Overflow is fine, just wrap
+            {
+                int hashCode = 41;
+                if (this.Seed != null)
+                    hashCode = hashCode * 59 + this.Seed.GetHashCode();
+                if (this.Resolution != null)
+                    hashCode = hashCode * 59 + this.Resolution.GetHashCode();
+                return hashCode;
+            }
+        }
+    }
+}