+++ Services/Image/V2/Model/ShowImageVariationTaskRequest.cs
@@ -1,0 +1,78 @@
+using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Linq;
+using System.Runtime.Serialization;
+
+using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+using HuaweiCloud.SDK.Core;
+
+namespace HuaweiCloud.SDK.Image.V2.Model
+{
+    /// <summary>
+    /// Request Object
+    /// </summary>
+    public class ShowImageVariationTaskRequest 
+    {
+
+        /// <summary>
+        /// 任务ID
+        /// </summary>
+        [SDKProperty("task_id", IsPath = true)]
+        [JsonProperty("task_id", NullValueHandling = NullValueHandling.Ignore)]
+        public string TaskId { get; set; }
+
+
+
+        /// <summary>
+        /// Get the string
+        /// </summary>
+        public override string ToString()
+        {
+            var sb = new StringBuilder();
+            sb.Append("class ShowImageVariationTaskRequest {\n");
+            sb.Append("  taskId: ").Append(TaskId).Append("\n");
+            sb.Append("}\n");
+            return sb.ToString();
+        }
+
+        /// <summary>
+        /// Returns true if objects are equal
+        /// </summary>
+        public override bool Equals(object input)
+        {
+            return this.Equals(input as ShowImageVariationTaskRequest);
+        }
+
+        /// <summary>
+        /// Returns true if objects are equal
+        /// </summary>
+        public bool Equals(ShowImageVariationTaskRequest input)
+        {
+            if (input == null)
+                return false;
+
+            return 
+                (
+                    this.TaskId == input.TaskId ||
+                    (this.TaskId != null &&
+                    this.TaskId.Equals(input.TaskId))
+                );
+        }
+
+        /// <summary>
+        /// Get hash code
+        /// </summary>
+        public override int GetHashCode()
+        {
+            unchecked // Overflow is fine, just wrap
+            {
+                int hashCode = 41;
+                if (this.TaskId != null)
+                    hashCode = hashCode * 59 + this.TaskId.GetHashCode();
+                return hashCode;
+            }
+        }
+    }
+}