+++ Services/Image/V2/Region/ImageRegion.cs
@@ -7,15 +7,20 @@ namespace HuaweiCloud.SDK.Image.V2
     public class ImageRegion
     {
         public static readonly Region CN_NORTH_4 = new Region("cn-north-4",
-            "https://image.cn-north-4.myhuaweicloud.com");
+            "https://image.cn-north-4.myhuaweicloud.com",
+            "https://image.cn-north-4.myhuaweicloud.cn");
         public static readonly Region CN_NORTH_1 = new Region("cn-north-1",
-            "https://image.cn-north-1.myhuaweicloud.com");
+            "https://image.cn-north-1.myhuaweicloud.com",
+            "https://image.cn-north-1.myhuaweicloud.cn");
         public static readonly Region AP_SOUTHEAST_1 = new Region("ap-southeast-1",
-            "https://image.ap-southeast-1.myhuaweicloud.com");
+            "https://image.ap-southeast-1.myhuaweicloud.com",
+            "https://image.ap-southeast-1.myhuaweicloud.cn");
         public static readonly Region AP_SOUTHEAST_3 = new Region("ap-southeast-3",
-            "https://image.ap-southeast-3.myhuaweicloud.com");
+            "https://image.ap-southeast-3.myhuaweicloud.com",
+            "https://image.ap-southeast-3.myhuaweicloud.cn");
         public static readonly Region CN_EAST_3 = new Region("cn-east-3",
-            "https://image.cn-east-3.myhuaweicloud.com");
+            "https://image.cn-east-3.myhuaweicloud.com",
+            "https://image.cn-east-3.myhuaweicloud.cn");
         
         private static readonly Dictionary<string, Region> StaticFields = new Dictionary<string, Region>()
         {